+++ src/main/java/it/denzosoft/jprolog/core/engine/BuiltInRegistry.java
@@ -42,10 +42,13 @@ public class BuiltInRegistry {
         // Unary predicates
         for (String p : new String[]{"assert", "asserta", "assertz", "retract", "retractall",
                 "current_predicate", "once", "ignore", "\\+", "writeln",
-                "get_char", "put_char", "get_code", "put_code", "close", "write",
+                "get_char", "put_char", "get_code", "put_code", "write",
                 "throw"}) {
             putArity(p, 1);
         }
+        // START_CHANGE: ISS-2025-0253 - close/1,2
+        putArity("close", 1, 2);
+        // END_CHANGE: ISS-2025-0253
         // START_CHANGE: ISS-2025-0203 - read/1 and read/2
         putArity("read", 1, 2);
         // END_CHANGE: ISS-2025-0203